requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# DB
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1

# Celery
celery==5.4.0
redis==5.0.4

# MinIO
minio==7.2.7

# PDF extraction
pdfminer.six==20231228
pypdf==4.2.0
pdf2image==1.17.0
numpy<2
opencv-python-headless==4.9.0.80

# OCR
pytesseract==0.3.10
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# File detection
python-magic==0.4.27

# Misc utils
PyYAML==6.0.1
httpx==0.27.0
structlog==24.1.0
python-dateutil==2.9.0
